Separate FormattingCodes from Display
